👌 Move console observable into domain
